nandi-ai-service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
pydantic-settings==2.2.1
openai==1.14.1
python-dotenv==1.0.1
slowapi==0.1.9
psutil==5.9.8
requests==2.31.0
celery==5.3.4
redis==5.0.0
python-multipart==0.0.6
httpx==0.25.0
tenacity==8.2.3 